+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Changelog
 - Support for ``--max-response-time`` in new-style stateful tests.
 - Support for targeted property-based testing in new-style stateful tests.
 - Support for ``--request-*`` CLI options in new-style stateful tests.
+- Support for ``--generation-*`` CLI options in new-style stateful tests.
 
 **Fixed**